Added auth to delete comment

+++ src/services/comment-service.js
@@ -28,17 +28,25 @@ class CommentService {
     }
   }
 
-  async delete(id) {
+  async delete(id, userId) {
     try {
       const comment = await this.commentRepository.get(id);
 
       if (comment) {
         var video = await this.videoRepository.get(comment.commentable);
-      }
-      if (video) {
-        video.comments.pull(comment.id);
-        await video.save();
-        return await comment.deleteOne();
+        if (video) {
+          if (userId === comment.userId || userId === video.userId) {
+            video.comments.pull(comment.id);
+            await video.save();
+            return await comment.deleteOne();
+          } else {
+            throw new Error("You are not authorized to perform this action");
+          }
+        } else {
+          throw new Error("Video not found");
+        }
+      } else {
+        throw new Error("Comment not found");
       }
     } catch (error) {
       console.error("Something went wrong at comment service layer: " + error);
